Article 4
Added a link for the new paper I wrote. The paper is an embedded Iframe that needed a in-page reference and a article tag to hold the paper. Google iframes were published from the Saas controls.

And rearranged the links on the sidebar.

Still todo:
need for media queries to create website reformats for different screen sizes.