Update About Page
Mobile Responsiveness to was added. Made the about page a flex box, that switches to column at smaller screens sizes.

Something, I did that kept me from fixing this issue was, putting absolute positioning on every element, which only looks good on some screen sizes ( the one it was meant for ).